requirements.txt
groq
gradio
faiss-cpu
sentence-transformers
numpy